Mobile friendly adding of links

+++ daemon.py
@@ -3975,18 +3975,33 @@ class PubServer(BaseHTTPRequestHandler):
 
             if fields.get('editedLinks'):
                 linksStr = fields['editedLinks']
+                if fields.get('newColLink'):
+                    if linksStr:
+                        if not linksStr.endswith('\n'):
+                            linksStr += '\n'
+                    linksStr += fields['newColLink'] + '\n'
                 try:
                     with open(linksFilename, 'w+') as linksFile:
                         linksFile.write(linksStr)
                 except OSError:
                     print('EX: _linksUpdate unable to write ' + linksFilename)
             else:
-                if os.path.isfile(linksFilename):
+                if fields.get('newColLink'):
+                    # the text area is empty but there is a new link added
+                    linksStr = fields['newColLink'] + '\n'
                     try:
-                        os.remove(linksFilename)
+                        with open(linksFilename, 'w+') as linksFile:
+                            linksFile.write(linksStr)
                     except OSError:
-                        print('EX: _linksUpdate unable to delete ' +
+                        print('EX: _linksUpdate unable to write ' +
                               linksFilename)
+                else:
+                    if os.path.isfile(linksFilename):
+                        try:
+                            os.remove(linksFilename)
+                        except OSError:
+                            print('EX: _linksUpdate unable to delete ' +
+                                  linksFilename)
 
             adminNickname = \
                 getConfigParam(baseDir, 'admin')
